Password Checker
This program was made using Python. The program asks the user to type
in a password that contains at least 8 characters and includes 1 letter
and 1 number. So far I have coded it to the point where the user will
get an error if the password dis less than 8 characters. I am still
working on the rest.